+++ CHANGELOG.md
@@ -3,7 +3,8 @@
 ### 2.8.1
 * improved the recovery of the information during the addition of a video;
 * updated for MeCms 2.16.1;
-* added tests for `AppController` class.
+* added tests for `AppController` class;
+* added tests for `isAuthorized()` method of all admin controllers.
 
 ### 2.8.0
 * the cells that act as widgets now have "Widgets" in the name, for the classes